Add support for [RequireExisting] attribute
When you add the [RequireExisting] attribute onto parameters, Protoinject will only satisfy the
parameter using an existing node in the hierarchy.  If you have an implementation which relies
on the parent or some other aspect of the hierarchy using another service, you can use this
attribute to ensure that your implementation does not incidentally create the implementation
if another service should be owning / requesting it.